+++ src/providers/dotnet/trx.types.ts
