+++ src/providers/javascript/istanbul.types.ts
