+++ tests/javascript/mocha/sample/.npmignore
