+++ tests/javascript/mocha/sample/tsconfig.json
