+++ tests/javascript/mocha/sample/tslint.json
